Working for basics, not clean or robust
But given we only need this once... we may not add anything else to it
unless we end up wanting to move epics over to sub issues.

If we do want that, we may just use https://github.com/agilesix/zh-to-gh instead of adding it into this tool,
though it shouldn't be hard to add to this tool.